Solved AoC2020 Day 2 Part 1

+++ aoc2020/aoc2020-kotlin/src/test/kotlin/de/havox_design/aoc2020/day02/Day02Test.kt
@@ -10,7 +10,7 @@ class Day02Test {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart1")
-    fun testProcessPart1(filename: String, expectedResult: Long) =
+    fun testProcessPart1(filename: String, expectedResult: Int) =
         PasswordPhilosophy(filename).processPart1().shouldBe(expectedResult)
 
     @ParameterizedTest
@@ -22,10 +22,10 @@ class Day02Test {
         @JvmStatic
         private fun getDataForTestProcessPart1(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2020/day02/day02part1sample1.txt", 2L),
-                Arguments.of("de/havox_design/aoc2020/day02/day02part1sample2.txt", 1L),
-                Arguments.of("de/havox_design/aoc2020/day02/day02part1sample3.txt", 0L),
-                Arguments.of("de/havox_design/aoc2020/day02/day02part1sample4.txt", 1L)
+                Arguments.of("de/havox_design/aoc2020/day02/day02part1sample1.txt", 2),
+                Arguments.of("de/havox_design/aoc2020/day02/day02part1sample2.txt", 1),
+                Arguments.of("de/havox_design/aoc2020/day02/day02part1sample3.txt", 0),
+                Arguments.of("de/havox_design/aoc2020/day02/day02part1sample4.txt", 1)
             )
 
         @JvmStatic
